Run Pronto in a GitHub action separate from the pull request
When the specs action is triggered by `pull_request` Pronto can not
post reports to GitHub because there isn't a token present (the action
only has read access to the repository).

Inspired by [this blog post][1] a second action is now triggered after
the specs finished. Since this action has write access to the repository
we must make sure that no code that is coming from the PR is executed.

We first download the artifact that contains the pull request number,
install the last released pronto version, check out the pull request
revision and finally run pronto.

There is an open issue in the `download-artifact` action to [add support
for workflow_run][2], until then the `github-script` snippet seems to be
the easiest way to download the artifact.

[1]: https://securitylab.github.com/research/github-actions-preventing-pwn-requests
[2]: https://github.com/actions/download-artifact/issues/60